requirements.txt
# Core dependencies
networkx>=3.1
pandas>=2.0.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
python-arango==7.1.0
jupyter>=1.0.0

# Graph Neural Network libraries
torch-geometric>=2.3.0
dgl>=1.1.0

# RAG-related dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langgraph==0.0.15
chromadb>=0.4.0
faiss-cpu>=1.7.4

# API and Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0

# Additional dependencies
matplotlib>=3.7.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
tqdm>=4.66.0

# Fixed versions for compatibility
marshmallow==3.15.0
arango-orm==0.7.2

aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
decorator==5.1.1
distro==1.9.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain-community==0.0.24
multidict==6.0.5
mypy-extensions==1.0.0
packaging==23.2
pillow==10.2.0
pydantic-core==2.16.3
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
scipy==1.15.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
tenacity==8.2.3
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4 